+++ test/forcing.jl
@@ -100,18 +100,19 @@ function check_fourier_band_forcing(forcing::FourierBandForcing, f::AbstractFila
     # Copy coefficients from vs_hat
     data_bs = @inferred BiotSavart.get_longrange_field_fourier(cache_bs)
     @test data_bs.state.quantity == :velocity
-    @test data_bs.state.smoothed == true  # this means that the field needs to be unsmoothed (by reverting Gaussian filter)
+    @test data_bs.state.smoothing_scale > 0  # this means that the field needs to be unsmoothed (by reverting Gaussian filter)
     vs_hat = map(Array, data_bs.field)  # GPU -> CPU copy in case `field` is a tuple of GPU arrays
     ks_grid = data_bs.wavenumbers
     SyntheticFields.from_fourier_grid!(vs_band, vs_hat, ks_grid)  # copy coefficients of smoothed velocity field
 
     # Revert Gaussian filter
-    α_ewald = cache_bs.params.α
+    σ_gaussian = data_bs.state.smoothing_scale
+    σ²_over_two = σ_gaussian^2 / 2
     let (; cs, qs, Δks) = vs_band
         for i in eachindex(cs, qs)
             k⃗ = Δks .* qs[i]
             k² = sum(abs2, k⃗)
-            cs[i] = cs[i] * exp(k² / (4 * α_ewald^2))  # deconvolve
+            cs[i] = cs[i] * exp(σ²_over_two * k²)  # deconvolve
         end
     end
 
@@ -152,7 +153,6 @@ end
 
 function get_energy_at_normalised_wavevector(cache::BiotSavart.BiotSavartCache, q⃗::NTuple{N}) where {N}
     Ls = cache.params.Ls
-    α_ewald = cache.params.α
     Δks = @. 2 * (π / Ls)
     k⃗ = q⃗ .* Δks  # unnormalised wavevector
     k² = sum(abs2, k⃗)
@@ -167,8 +167,9 @@ function get_energy_at_normalised_wavevector(cache::BiotSavart.BiotSavartCache, 
     @assert all(!isnothing, inds_pos)
 
     # Unsmoothed velocity at wavevector k⃗
-    @assert state.quantity == :velocity && state.smoothed == true
-    v̂ = Vec3(map(u -> u[inds_pos...], field)) * exp(k² / (4 * α_ewald^2))
+    @assert state.quantity == :velocity
+    σ = state.smoothing_scale
+    v̂ = Vec3(map(u -> u[inds_pos...], field)) * exp(σ^2 * k² / 2)
 
     # Energy at wavevectors k⃗ and -k⃗ (assuming Hermitian symmetry) -> hence no division by 2!
     Ek = sum(abs2, v̂)
@@ -193,7 +194,7 @@ end
     vs_hat = data.field
     state = data.state
     @test state.quantity == :velocity
-    @test state.smoothed == true
+    @test state.smoothing_scale > 0
 
     # Initialise random normal fluid velocity in Fourier space
     rng = StableRNG(42)